feat: range complete

+++ keepmeup/main.py
@@ -4,10 +4,10 @@ import random
 import argparse
 import datetime
 
-def dontsleep(lunch_mode=False, time_input=0):
+def dontsleep(lunch_mode=False, time_input=0,time_range=None):
 
     def pressingbuttons():
-        down, up = random.randint(1, 15), random.randint(1, 15)
+        down, up = random.randint(1, 100), random.randint(1, 100)
         pyautogui.press('volumedown')
         print(f'volumedown - rest for ({down}) sec(s)')
         time.sleep(down)
@@ -34,12 +34,19 @@ def dontsleep(lunch_mode=False, time_input=0):
             if timer > 0:
                 print(f"Logic will run for another {timer} seconds")
         print(f"Termining program :(")
+
+    elif time_range is not None:
+        start,end = time_range
+        while True:
+            current_time = datetime.datetime.now().time()
+            if current_time >= datetime.time(int(start),0) and current_time < datetime.time(int(end),0):
+                pressingbuttons()
     
     else:
         pressingbuttons()
 
-def KeepUI(lunch_mode=False,time_input=0):
-    p2 = multiprocessing.Process(target=dontsleep(lunch_mode, time_input))
+def KeepUI(lunch_mode=False,time_input=0,time_range=None):
+    p2 = multiprocessing.Process(target=dontsleep(lunch_mode, time_input,time_range))
     p2.start()
     return p2
 
@@ -49,6 +56,7 @@ def main():
         description='Wakey wakey :)')
     parser.add_argument('--lunch', help='times out during lunch', action=argparse.BooleanOptionalAction)
     parser.add_argument('--time', help='how long you want to stay up for (in minutes) (will always minus 5)', type=int)
+    parser.add_argument('--range', help='The time period to keepmeup', type=str)
     args = parser.parse_args()
 
     if args.lunch:
@@ -66,6 +74,20 @@ def main():
             print("Press ctrl+c to kill me")
             print(f"Program will stop in roughly {args.time-5} minute(s)")
             p1 = KeepUI(time_input=args.time)
+            p1.terminate()
+    elif args.range:
+        range = len(args.range.split(','))
+        if range == 2: 
+            morning, afternoon = args.range.split(',')
+            morning_range = morning.split('-')
+            p1 = KeepUI(time_range=morning_range)
+            p1.terminate()
+            afternoon_range = afternoon.split('-')
+            p1 = KeepUI(time_range=afternoon_range)
+            p1.terminate()
+        if range == 1:
+            total_range = args.range.split('-')
+            p1 = KeepUI(time_range=total_range)
             p1.terminate()
     else:
         print("Press ctrl+c to kill me")
